Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/css-timing-1/frames-timing-functions-syntax-expected.txt


PASS The number of frames must be a positive integer greater than 1, or we fallback to the previously-set easing 
FAIL The serialization of frames is 'frames(n)', n is the number of frames assert_equals: expected "frames(2)" but got "ease"
